refactoring : quiz / alarm / recentlyissue 폴더 구조 수정
