+++ src/app/alarm/page.tsx
@@ -1,11 +1,9 @@
-
-import AlarmItem from '@/components/alarm/AlarmItem';
-
+import AlarmItem from '@/app/alarm/(alarm)/AlarmItem';
 
 const AlarmPage = () => {
   return (
-    <div className='h-full flex flex-col sm:mb-0'>
-      <AlarmItem/>
+    <div className="h-full flex flex-col sm:mb-0">
+      <AlarmItem />
     </div>
   );
 };
