+++ src/app/newissue/(recentlyIssue)/Index.tsx
@@ -3,7 +3,7 @@
 import React, { useEffect } from 'react';
 import BestNews from './BestNews';
 import CategoryNews from './CategoryNews';
-import Container from '../common/Container';
+import Container from '../../../components/common/Container';
 
 const RecentlyIssueIndex = () => {
   /**
