+++ src/app/quiz/(quiz)/QuizBanner.tsx
