+++ src/app/quiz/start/page.tsx
@@ -1,4 +1,4 @@
-import QuizeIndex from '@/components/quiz/QuizeIndex';
+import QuizeIndex from '@/app/quiz/(quiz)/QuizeIndex';
 
 const QuizePage = () => {
   return (
